requirements.txt
pulp
pandas
streamlit
plotly
matplotlib.pyplot
